firestore: rewrite Client.getAll
Simplify the method that does a batch get of documents. The new
implementation no longer needs a separate struct type to store
results. Also, it always reads the stream to the end, where the
earlier one might exit early with an error. That early exit would have
required that we make a new cancelable context and cancel it on exit,
because gRPC streams need to be read to the end or have their context
canceled in order to be cleaned up.  That would have added a little
more complexity.

We also no longer check for a particular error, but we never really
needed to anyway because it would be a server bug.

Change-Id: I3cd94b3def6a5d6f9fd118aed85ef1336a177e67
Reviewed-on: https://code-review.googlesource.com/25750
Reviewed-by: Jean de Klerk <deklerk@google.com>