firestore: return same number of snapshots as doc refs
Firestore GetAll should return the same number of document
snapshots as the number of document references that was
specified by the user. If the same document reference was
specified multiple times, the method returns one document
snapshot for each document reference specified by the user.
The returned document snapshots have the same order as the
order of the input document references.

Fixes #1437

Change-Id: I4e4089d2a6730c6cebffcbfe57ad5a578380f4a4
Reviewed-on: https://code-review.googlesource.com/c/gocloud/+/41810
Reviewed-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Jean de Klerk <deklerk@google.com>